requirements.txt
#https://github.com/nik0811/conu
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
google-auth==1.34.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
kubernetes>=9
MarkupSafe==2.0.1
oauthlib==3.1.1
openapi-codec==1.3.2
Pillow==8.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
websocket-client==1.1.0
whitenoise==5.3.0
python-dotenv
